requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
elementpath==5.0.4
ibm-cloud-sdk-core==3.24.2
ibm-watson==10.0.0
idna==3.10
lxml==6.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
paho-mqtt==1.6.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
six==1.17.0
urllib3==2.5.0
websocket-client==1.8.0
wheel==0.45.1
xmlschema==4.1.0
fastapi
uvicorn[standard]
docker
asyncio